requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
bcrypt
PyJWT
pillow
torch
torchvision
anthropic
